fix: allow the snap only on Shibuya, moved snapUtils to the snap module
