+++ src/boot/api.ts
@@ -17,7 +17,7 @@ import { useMetaExtensions } from 'src/hooks/useMetaExtensions';
 import { computed, ref, watchPostEffect } from 'vue';
 import Web3 from 'web3';
 import { SupportWallet, supportWalletObj } from 'src/config/wallets';
-import { initiatePolkdatodSnap } from 'src/hooks/helper/snapUtils';
+import { initiatePolkdatodSnap } from 'src/modules/snap';
 import { initPolkadotSnap } from '@chainsafe/metamask-polkadot-adapter';
 
 let $api: ApiPromise | undefined;
